+++ common/src/main/java/com/sonicether/soundphysics/SoundPhysicsMod.java
@@ -1,5 +1,6 @@
 package com.sonicether.soundphysics;
 
+import com.sonicether.soundphysics.config.OcclusionConfig;
 import com.sonicether.soundphysics.config.ReflectivityConfig;
 import com.sonicether.soundphysics.config.SoundPhysicsConfig;
 import net.minecraft.client.Minecraft;
@@ -10,6 +11,7 @@ public abstract class SoundPhysicsMod {
 
     public static SoundPhysicsConfig CONFIG;
     public static ReflectivityConfig REFLECTIVITY_CONFIG;
+    public static OcclusionConfig OCCLUSION_CONFIG;
 
     public void init() {
 
@@ -17,6 +19,7 @@ public abstract class SoundPhysicsMod {
         CONFIG.reload();
 
         REFLECTIVITY_CONFIG = new ReflectivityConfig(Minecraft.getInstance().gameDirectory.toPath().resolve("config").resolve(MODID).resolve("reflectivity.properties"));
+        OCCLUSION_CONFIG = new OcclusionConfig(Minecraft.getInstance().gameDirectory.toPath().resolve("config").resolve(MODID).resolve("occlusion.properties"));
     }
 
     public abstract SoundPhysicsConfig createConfig();